Descriptor buffers, eliminates descriptor sets and pools
Descriptor buffers are raw byte buffers that can be used to store descriptors instead of using pools and sets. Works well with the bindless setup that I did from before.
Now I need to make it easier to allocate and free descriptors with this setup to make it more usable.