Normal mapping
Implemented normal mapping, lighting is calculated in world space, so the sampled normal is transformed from tangent to world space in the fragment shader.
Changed loading of textures to have the correct format, base color should be in SRGB while normal and metallic roughness textures should be in UNORM (linear).
Added mikkt for tangent calculation, should they be missing from a model.