Render to cubemap directly instead of intermediate render target, separated images from image views, added image views and layout transitions for parts of an image
Separated the concept of images and image views in the vulkan backend, which now makes it possible to create multiple image views into the same physical image.
Image layout transitions can now be performed on individual mips or layers within the image.
Generating the cubemap from an hdr map will no longer render to an intermediate render target and then do a copy, but instead render to the cubemap faces directly.