Temporary way of chaining additional textures onto one primary texture resource
This is used for the IBL implementation to chain the irradiance cubemap onto the primary hdr cubemap.
I kinda want a better way of doing this, maybe kinda like vulkan how it has a pNext pointer to additional data, and specifies in the first 4 or so bytes what it is.