Aligned descriptor allocations, fixed UBO descriptors and bindings
UBOs have a fixed binding now, so that I do not have to pass in a ubo index for each ubo I want to access. That makes it easier to use UBOs that are relevant for many different render passes/shaders.
Changed the descriptor buffer to have a special case if the descriptor type is UNIFORM_BUFFER. This is a temporary solution, the descriptor buffer abstraction I will come back to later and redo.
Renamed EquirectangularToCube.vert to Cube.vert.