Some BRDF and clearcoat fixes
Fixed an oversight in the base BRDF specular schlick term, where the dot product of the normal and the view vector instead of the light and half vector was used.
Fixed the evaluation for clearcoats in direct lighting.
Fixed the evaluation for clearcoats in indirect lighting.
Tested with white furnace for image-based lighting.
Direct lighting white furnace testing coming next.