Changed camera data slightly to give more freedom to the renderer itself, fixed minimizing
Camera no longer responsible for the projection matrix, the projection matrix will now be updated in the Renderer::BeginFrame.
The renderer now keeps track of the near and far planes for the cameras, as well as the aspect ratio for the projection matrix.
The Camera class only keeps track of the vertical FOV from now.
Fixed minimizing which crashed before.