Analytic area lights using Linearly Transformed Cosines
This is still WIP, its a first implementation that seems to work, but needs to be tested properly after this.
Also need to render the actual rectangular area light to make it visible in the world.
Changed the way that a mesh's name is determined while loading it from a GLTF. Should now always create unique names.
Added AreaLight entity type.
Renderer::SubmitPointlight function currently does nothing.
Added a batch file to copy the shaders from the assets folder to the binary builds.